Remove SWIG from the build process - unfortunately, it's slightly broken. Also, it made me slightly uneasy that there was no way to prevent swig from putting one of its generated files in sorce directory. So, I've just checked in the two major files that get generated by SWIG, so that the make rule that runs it never gets invoked.
One of the reasons for doing this is that swig generates slightly
broken code when the -exportall (which does perl module exports
correctly) arugment is given. A very minor amount of manual tweaking
of the generated .pm file can fix this problem. So, the checked in
copy of event.pm has these tweaks applied.

As a result of all of this, exports work correctly in the event perl
module, so the hacky practice of putting your program in the event
namespace is no longer necessary.