Add generation of per-project email lists, as per Dave's request.  The lists are stored on users:/etc/mail/lists. For example, you can send email to ron-users@users.emulab.net. An alias entry is added (and newaliases run) if there is no alias in /etc/mail/aliases (by the proxy of course). There are two new options to genelists (on boss):
	"Use the -a option to generate lists for all projects.\n".
	"Use the -n option to generate lists for a new user.\n";

With no options, generate the all users and active users lists
(renamed to emulab-users and emulab-active-users). With the -n option
(used by mkacct) regen the lists for all the projects the user is a
member of.

It would be nice to archive the email, but that requires a publically
readable directory and a u+S archive file; the mailer daemon runs as
user daemon, and the project tree is 770, so it cannot write the
archive file. At some point we will have to go to majordomo or spam
filtering, when the first list is spamm'ed. Sigh.