Support for a "reset" command to close existing connection.
Also fixed bug where changing packet size would change an explicitly
specified packet rate because we only tracked the packet interval.  E.g.:
	set packetSize_=100
	set rate_=100
would result in a packet rate of 100Kbps, but then doing:
	set packetSize_=1000
would result in a packet rate of 1000Kbps even though the rate was not
explicitly changed.  Now we actually keep track of whether the rate was
explicitly set and, if so, change the interval to maintain the rate.