Add tunnels table, loaded by assign_wrapper. This tables lists what tunnels need to be created on each node; will be returned as a new tmcd command.
Add nodeipportnum table, which is a hacky way to allocate non
overlapping port numbers on physical nodes (when multiple virtual
nodes are mapped to them). Used in assign_wrapper as a post pass over
the trafgens table and the tunnels table.

Add target_vname to virt_trafgens table as an aid in post passing.