Add support for link costs for routing algorithm. These are specified in the NS file in one of two ways:
$ns cost $nodeA $nodeB 2
or
$link0 cost 3
(The latter form can also be used for LANs)

This information is stored in the database in the cost column of the
virt_lans table.  It's important to note that cost data is _not_
divided between the two sides of the link, as things like delay and
rdelay are. Since there are no physical devices (ie.  delay nodes)
node that this configuration needs to be split up between, doing so
would just be unnecessary work.