- Changed internal representation to accommodate >1 vnodes per pnode.
- Added '-m' option to allow the user to specify
the maximum number of vnodes mappable onto each pnode.
(if -m is not specified, only one vnode is allowed per pnode)

- Added ability to "fix" a permanent mapping from a vnode to a pnode.
When specifying a vnode name, use the name of a pnode, with a "*"
prepended. This will be useful to fix one node as "The Testbed Node."

See comments at beginning of source for full documentation.