Added "hole"ability in the physical matrix; placing a "-1" anywhere in the latency or bandwidth matrices will result in huge penalties when that connection is used in the mapping.
Also added neat (possibly novel?)
"mutation spike" which, when the standard deviation of the population gets
beneath a certain value, spikes the mutation rate from 1% to 50% for a generation,
thus keeping the deviation (and the population diversity) up.
Results from testing look promising.. in my "standard" 6-node mapping,
the solver finds the "best" solution in 95% of the runs.
(vs about 50% or less the old way.)