Belated checkin of a lot of new pcapper features. Highlights:
Event system:

When compiled with -DEVENTSYS, can be made to wait for time to start
in an experiment before starting to count packets. Times are also
reported relative to experiment time start. Use the '-e' flag to
enable.

SUID support:

If compiled with -DDROPROOT, and it seems that pcapper was started
setuid root (euid == 0, and ruid != 0), drops root permissions after
opening BPF (or raw socket in Linux.) This allows it get permissions
to read packets, but still be killed by the user. This is particularly
useful with the testbed's program objects.

New command-line options:

-s: Print out packet counts to stdout, in addition to listening
        on a socket ('-f -' does the same thing, too)
-p: Count only payload sizes, not header sizes
-e: Wait for event system time to start in pid/eid
-z: Don't count zero-length packets in the packet counts (useful
        with '-p')

Also, now creates a GNUmakefile the _right_ way.