Added a GNUmakefile for the tools directory.
First, gives us a handy way to build all the tools, if there
is ever more than one.

Second, it's a workaround for a really annoying problem with
configure. Since there was nothing in the tools/ directory itself,
it wasn't getting created, so configure could not make
tools/pcapper (since the parent directory didn't exist.)