Ug, finish up the verify user changes. I can't believe I forgot to do this last part! Senility is setting in. Anyway, change join project to not send the email to the projleader when the user is new; let that happen out of the verify page when the newuser does the verification. Basically, duplicate some code to generate the email message.
XXX: I have not fixed up approveuser.php3 to check to make sure that
the user has been verified. Basically, for this to happen the proj
leader would have to generate the URL by hand, and thats not likely,
and not really a dangerous thing anyway. However, it is confusing.