First useful version of idlecheck. Current settings:  - 4 packets/hour avg over last 24 hours (100pkts) is activity  - use of a tty in last 24 hours is activity.  - Latest report from slothd must be within last 120 minutes.
Caveats:
 - Shows all expts that are not known to be active. This includes
   expts that we don't have any information on. I plan to fix this so that
   expts that are reporting are very different from those without reports.
 - Because of the above, we're quite conservative in marking nodes
   active. When the above gets fixed, this will be better.
 - Displays unswappability. We may in some cases still want to harass
   owners of unswappable expts, since they can mark it arbitrarily.
 - Currently takes ~10 seconds to run. I plan to run the queries
   through explain and try to optimize them a little better, or even
   simplify where possible.

The output of this script is what will eventually get used in idle
view of the web page. Unless optimization makes this script run in
less than 1 second, we'll most likely need some sort of caching of
these results before being able to use them in the web page. Perhaps
stashing them in the db with the time the idleness was detected, so it
is easy to get to.