First, fixed a bug reported by Leigh. Wasn't checking for swapped-out experiments early enough.
Then, switched to using the new TBWideareaNodeID call so that we can
report links that go between RON nodes and the testbed properly.

Also, made a new message, seperate from "No data available" for vnodes
that are on the same physical node, or that seem to be at the same
site.