Two changes:
First, remove references to the IP columns, since those are now gone.

Second, we now support a "magic node" that is assumed to be equivalent
to the TBDB_WIDEAREA_LOCALNODE (ususally boss). Updates for the magic
node are also applied to the localnode. Right now, this magic node is
hardcoded in this script, on the assumption that no one but us will
ever use this script. But, we could move it to a configure variable or
even the DB if we need to.