The big key changes ... Deprecate the two pubkey slots in the users table and create a new table to hold user_pubkeys, indexed by the comment field of the key. Change mkacct to insert newly created Emulab keys into that table, and to regen the users authorized_keys file from the DB. Users should no longer edit their own authorized_keys file or the changes will be lost (I put a comment in their files).
Change the three pages that deal with keys. join/new project can now
take a file of multiple keys; each is inserted. Moved the key stuff
that was in the update user info page into a new pubkeys page that
allows users to add/sub keys easily. New key additions are password
protected.

Unrelated change: Add an audit mode to mkacct to log its output and
send it to the tblogs email. Previously, warnings and errors tended to
get lost.