Add swap request tracking. Two new columns in expt table (swap_requests and last_swap_req) track the count and last time a request was sent.
The counters all get reset by idlecheck every time it runs unless it
detects you as inactive. Then when you go to send a swap request, it
will show you the info about how many have been sent and when the last
one was, and will let you know if the swap request will also go to the
proj leader if you send it. When you send the request, the email
mentions how many reminders you've been sent already, and only gets
cc'd to proj head on the third message (or more). Then it updates the
count and the timestamp of the last request sent.