Last part of the recent widearea changes. Now that widearea nodes are reporting in isalive status, and since we are returning a "need to update" flag in the message, and since we are now initiating widearea account updates by setting the flag for each node, and since tmcd decrements that flag when the node picks up accounts, it is a simple matter to do update for remote nodes! Simply, bump the flag in this script, and then wait a little while to see whose flag does not change back! Report those in the email message. Works very slick.
This mechanism is going to be generalized some to handle other kinds
of updates. Either via the isalive return value, or via a specialized
TMCD call to return the list of things that need to be updated. The
next most needed update mechanism we need is for the testbed software,
followed closely by user software.