Fix for the problem Eric reported for this topology:
	uav <---> distrib <---> display
         ^           ^             ^
         +-----------+-------------+---> namer

where tmcd was returning multiple aliases since there were multiple links
between nodes. Thats been a bug for long time, but its very rare to have
two links between the same two nodes, and so I never bothered to deal with
it. I have "fixed" it, but the fix is very unsatisfying since the choice it
makes is not consistent. For example, on distrib the ideal choice is to
create aliases for the lan links, but instead I get a mix; one alias is for
the duplex link to uav, another to the lan link to display. To make it
select a consistent set, it will take even more hackery in tmcd, and I am
not in the mood for it right now, given how rarely this happens. At least
the host file is now "correct."