Several small but much-needed changes to idle detection. Soon I'll give this stuff a major overhaul when we move to the model where we have data on when each node was last "touched" or actively used. (Most of these changes will still be relevant.)
1. Fix a bug in idlecheck that we didn't really thing much about. It
turns out that the WanSpread people have had a 17 node expt idle for
over three weeks, we didn't detect it because the nodes were running
ospf and generating lots of network traffic on the exptl. net. We now
ignore the exptl. network traffic as a source of activity if they have
automatic ospf routing happening. We also ignore nodes that have any
trafgen endpoints.

2. After the changes in idlecheck, I updated idle view to correctly
handle a new flag idlecheck outputs to let us know that an expt has
ospf running and may be falsely inactive because we ignored net
traffic. It will probably be very rare that an active expt that is
running ospf will have only network activity and no tty activity, but
it is a potential source of false positives. I also did some random
libifying. There were some hard coded references to emulab.net that I
fixed to properly use the variables from the defs file.

3. I also updated request_swapexp.php3. It now inlcudes in the email
message a blurb about automatic swapping, depending on whether or not
they're marked swappable. (If swappable, it says "This experiment is
marked as swappable, so it may be automatically swapped out by
Emulab.Net or its operational staff." and if unswappable, it says
"This experiment has not been marked swappable, so it will not be
automatically swapped out.") It also has a reference to the Node Usage
Policy and gives the URL. So we now give them fair warning about
potentially getting swapped out and what our policies are.