Redefined Windows usage...
Core realization:

Since the installer does an excellent job of associating the tbacl filetype to tiptunnel
properly *in the shell*, users can save their acl files to their desktop (or whatever) and
double-click them to connect. This works perfectly smoothly, and obviates the need to mess around with
MIME types.

So, this is the new "official method" for windows.


Plus, we can spin it as more convenient than the web method
(e.g. put all your acl files somewhere, so you can quickly connect
 to your nodes without having to log in or navigate to them.) .

Barring any more unpleasant surprises, with this usage,
TipTunnel should be good for the tutorial.