Add tbshutdown daemon. Started from rc.setup right after we tell the testbed we've rebooted. Immediately daemonizes, then sets a signal handler and sleeps forever. When it gets a signal to quit (on shutdown, reboot or going single-user), it sends a state transition to SHUTDOWN to the testbed. It installs into /etc/testbed/ .
Important note:
If you're going to use this in an image that runs the NORMAL state machine
(instead of NORMALv1, MINIMAL, etc.), the SHUTDOWN state is called
REBOOTING in that mode. Just change the appropriate line in tbshutdown,
but don't commit the change, since further use of the NORMAL mode in new
images is deprecated. Alternatively, you could change the image over to
NORMALv1 by renaming all REBOOTING tags to SHUTDOWN and changing REBOOTED
to TBSETUP.