Server: back to using a condvar since they seem to be fixed.
Server: make file readsize independent of burstsize (previously
readsize had to be a divisor of burstsize).  A subtle side-effect
is that the dynamic burst rate is recalcluated at the conslusion
of every burst instead of after every readsize count of blocks has
been sent (less than a burst)  This just seems to be more logical.

Client: add "-T DOS-type" option to tell frisbee, when in slice
mode, to set the type of the slice in the DOS partition table.
This is useful if you are dropping say a BSD filesystem into
an unused slice, you don't have to go back later and set this
with fdisk.  Considered making this info part of the image
itself (recorded by imagezip when creating a slice image),
but decided against it.