Two new routines. TBNodeBootReset() resets the startup state for a node. Used in new tbrestart code for replaying experiments. It remains to be seen if this is a workable approach.
TBNodeStateWait() is really WaitTillAlive, which I need in several new
spots now. Its not as general purpose as it seems though, since there
are only a couple of terminal states (isup) that you can actually wait
for by querying the DB. But, I'm loathe to add any more event code to
the system.