Didn't have routes to other interfaces of directly connected nodes. For e.g.
                    1.1    1.2
                   A---------B
                3.1\        /2.1
                    \      /
                  3.2\   /2.2
                       C

In the above topology, there were no routes on C to ip addresses 1.1 and 1.2
because the code skipped adding routes to directly connected nodes.

This fixes the problem. We still don't add routes to the ip address on the
directly connected interface. For e.g., On C, we don't add a route for 3.1
We just add host routes for 1.1 and 1.2

C add-route -host 1.1 3.1
C add-route -host 1.2 2.1

Now, we should have every experimental interface pingable and reachable via
the shortest path.