Move addpubkey from the utils to account directory. Note that I copied the RCS control file in the repository so the history is left intact.
Two new modes, which used to be in the old mkacct. There is an init
mode, which is used on new users to create the initial pub key. There
is also a write mode, which is used regenerate the authkeys files for
people after they add/delete keys via the web interface. Used to be
that addpubkey wold add the key to the DB, but mkacct would deal with
creating the authkeys files. All this functionality is now localized
in this one script.