Add tb-set-uselinkdelays to switch between normal delay nodes and link delays setup on the nodes.
Add tb-set-forcelinkdelays to force a linkdelay (when uselinkdelays is
on of course) to be inserted, even if the link was unshaped. This
allows it to be controlled later, say via the event system. This is
nice, cause in the current system an unshaped link cannot be later
shaped.

Note that these are currently require FBSD-47.