Remove "be patient message" since I had added a redirect and that message screws up the redirect!
Note: Need to remove the silly "please be patient" messages. I've been
removing those and adding redirects since these messges are never seen
(the page is not drawn until the closing </table> is sent over). I
have a little javascript thing that throws up a be patient box, that
disappears when the page is finished. Need to get that installed some
day. Worked fine on opera and explorer, but on netscape it looked
totally stupid. Netscape, what else is new.