Add three new control mechanisms to the front end to control the use of emulated and linkdelays on a per-link/lan basis (although you cannot make a lan emulated at this time).
	tb-set-link-uselinkdelay $link0 1
	tb-set-lan-uselinkdelay $lan0 1
	tb-set-link-emulated $link0 1

Added slots to the virt_lans table, used in assign_wrapper to control
these features on a per link/lan basis. Note, the global options are
still available to prefer/force linkdelays for an entire experiment.
When setting uselinkdelays on a per link/basis, if the link would not
otherwise be shaped, you are requesting that a linkdelay be inserted
anyway (in otherwords, its the equiv of tb-set-forcelinkdelays on an
individual basis).