Add target_pid,target_gid arguments to joinproject page so that we can send links that cause those fields to be filled in for people. Add those links to email generated by newproject and newgroup pages so that they can be saved by the leaders.
Add a bit of referrer magic to login page. When clicked from the join
or new project pages, tell login page to pass along the referrer page
so that when login is complete, user is zapped back to the original
page. This is especially nice when combined with the above change to
joinproject, whereby we send along the target pid/gid, but the user
has not yet logged in and remembers to do so via the link at the top
of the page.