Massive reworking of our structure for defs files.
Before:
The main defs file (ie for configure) had an entry for WWWDEFS that
pointed to a <@WWWDEFS@>-defs.php3 file in the www/ directory. The www
defs file loaded some values about web pages, URLs, and some web
configuration parameters.

Problem:
Anything that was only in the www defs file was not accessible in the rest
of the universe (ie perl, C, and any other non-web-page scripts). For
instance, you couldn't have a perl script send an email to a user with a
link to the web site.

Solution:
Nuke all the www-defs files, move any important values into the main
configure, and change the web defs infrastructure to respect that. This
also meant adding about 3 lines each to all of the configure defs files.
(There really are about 10 new values you can change in your defs file,
but in almost all cases, the default values are the right thing.)

Upgrading:
External sites will need to move a few variables from their www-defs file
into their configure defs file. The example file should make it pretty
obvious. They may also want to customize some of the other vars that are
mentioned in configure.in and www/defs.php3.in .