Added 'sitevariables' table.
This is the table prompted by Mac's email yesterday, which
can contain site-specific configuration options.

'default' _must_ be set for each row, and will be used if
'value' is NULL.

'description' _should_ be set for each row.

The table doesn't have a primary key, but is indexed to
32 characters on the "name" column, so it should be plenty efficient.

libdb functions and web interface forthcoming...